Move all uts under *_parser.rs to tests/
- Merge duplicate or overlapped tests
