+++ sgl-router/src/tool_parser/parsers/deepseek_parser.rs
@@ -250,53 +250,3 @@ impl ToolParser for DeepSeekParser {
         self.has_tool_markers(text)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_deepseek_single_tool() {
-        let parser = DeepSeekParser::new();
-        let input = r#"Some text
-<｜tool▁calls▁begin｜><｜tool▁call▁begin｜>function<｜tool▁sep｜>get_weather
-```json
-{"location": "Tokyo", "units": "celsius"}
-```<｜tool▁call▁end｜><｜tool▁calls▁end｜>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Tokyo"));
-        assert_eq!(normal_text, "Some text\n")
-    }
-
-    #[tokio::test]
-    async fn test_parse_deepseek_multiple_tools() {
-        let parser = DeepSeekParser::new();
-        let input = r#"<｜tool▁calls▁begin｜><｜tool▁call▁begin｜>function<｜tool▁sep｜>get_weather
-```json
-{"location": "Tokyo"}
-```<｜tool▁call▁end｜>
-<｜tool▁call▁begin｜>function<｜tool▁sep｜>get_weather
-```json
-{"location": "Paris"}
-```<｜tool▁call▁end｜><｜tool▁calls▁end｜>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert_eq!(tools[1].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Tokyo"));
-        assert!(tools[1].function.arguments.contains("Paris"));
-        assert_eq!(normal_text, "");
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = DeepSeekParser::new();
-        assert!(parser.detect_format("<｜tool▁calls▁begin｜>"));
-        assert!(!parser.detect_format("plain text"));
-        assert!(!parser.detect_format("[TOOL_CALLS]"));
-    }
-}
