+++ sgl-router/src/tool_parser/parsers/glm4_moe_parser.rs
@@ -240,80 +240,3 @@ impl ToolParser for Glm4MoeParser {
         self.has_tool_markers(text)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_glm4_single_tool() {
-        let parser = Glm4MoeParser::new();
-        let input = r#"Some text
-<tool_call>get_weather
-<arg_key>city</arg_key>
-<arg_value>Beijing</arg_value>
-<arg_key>date</arg_key>
-<arg_value>2024-06-27</arg_value>
-</tool_call>More text"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Beijing"));
-        assert!(tools[0].function.arguments.contains("2024-06-27"));
-        assert_eq!(normal_text, "Some text\n");
-    }
-
-    #[tokio::test]
-    async fn test_parse_glm4_multiple_tools() {
-        let parser = Glm4MoeParser::new();
-        let input = r#"<tool_call>get_weather
-<arg_key>city</arg_key>
-<arg_value>Beijing</arg_value>
-</tool_call>
-<tool_call>get_weather
-<arg_key>city</arg_key>
-<arg_value>Shanghai</arg_value>
-</tool_call>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert_eq!(tools[1].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Beijing"));
-        assert!(tools[1].function.arguments.contains("Shanghai"));
-        assert_eq!(normal_text, "");
-    }
-
-    #[tokio::test]
-    async fn test_parse_glm4_mixed_types() {
-        let parser = Glm4MoeParser::new();
-        let input = r#"<tool_call>process_data
-<arg_key>count</arg_key>
-<arg_value>42</arg_value>
-<arg_key>active</arg_key>
-<arg_value>true</arg_value>
-<arg_key>name</arg_key>
-<arg_value>test</arg_value>
-</tool_call>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(normal_text, ""); // Pure tool call, no normal text
-        assert_eq!(tools[0].function.name, "process_data");
-
-        // Parse arguments to check types
-        let args: serde_json::Value = serde_json::from_str(&tools[0].function.arguments).unwrap();
-        assert_eq!(args["count"], 42);
-        assert_eq!(args["active"], true);
-        assert_eq!(args["name"], "test");
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = Glm4MoeParser::new();
-        assert!(parser.detect_format("<tool_call>"));
-        assert!(!parser.detect_format("plain text"));
-        assert!(!parser.detect_format("[TOOL_CALLS]"));
-    }
-}
