+++ sgl-router/src/tool_parser/parsers/gpt_oss_parser.rs
@@ -227,66 +227,3 @@ impl ToolParser for GptOssParser {
         self.has_tool_markers(text) || text.contains("<|channel|>commentary")
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_gpt_oss_single_tool() {
-        let parser = GptOssParser::new();
-        let input = r#"Some text
-<|channel|>commentary to=functions.get_weather<|constrain|>json<|message|>{"location": "San Francisco"}<|call|>
-More text"#;
-
-        let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("San Francisco"));
-    }
-
-    #[tokio::test]
-    async fn test_parse_gpt_oss_multiple_tools() {
-        let parser = GptOssParser::new();
-        let input = r#"<|channel|>commentary to=functions.get_weather<|constrain|>json<|message|>{"location": "Paris"}<|call|>commentary
-<|channel|>commentary to=functions.search<|constrain|>json<|message|>{"query": "Paris tourism"}<|call|>"#;
-
-        let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert_eq!(tools[1].function.name, "search");
-        assert!(tools[0].function.arguments.contains("Paris"));
-        assert!(tools[1].function.arguments.contains("Paris tourism"));
-    }
-
-    #[tokio::test]
-    async fn test_parse_gpt_oss_with_prefix() {
-        let parser = GptOssParser::new();
-        let input = r#"<|start|>assistant<|channel|>commentary to=functions.test<|constrain|>json<|message|>{"key": "value"}<|call|>"#;
-
-        let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "test");
-    }
-
-    #[tokio::test]
-    async fn test_parse_gpt_oss_empty_args() {
-        let parser = GptOssParser::new();
-        let input =
-            r#"<|channel|>commentary to=functions.get_time<|constrain|>json<|message|>{}<|call|>"#;
-
-        let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_time");
-        assert_eq!(tools[0].function.arguments, "{}");
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = GptOssParser::new();
-        assert!(parser.detect_format("<|channel|>commentary to="));
-        assert!(parser.detect_format("<|channel|>commentary"));
-        assert!(!parser.detect_format("plain text"));
-        assert!(!parser.detect_format("[TOOL_CALLS]"));
-    }
-}
