+++ sgl-router/src/tool_parser/parsers/kimik2_parser.rs
@@ -247,60 +247,3 @@ impl ToolParser for KimiK2Parser {
         self.has_tool_markers(text) || text.contains("<|tool_call_begin|>")
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_kimi_single_tool() {
-        let parser = KimiK2Parser::new();
-        let input = r#"Some text
-<|tool_calls_section_begin|>
-<|tool_call_begin|>functions.get_weather:0<|tool_call_argument_begin|>{"location": "Tokyo", "units": "celsius"}<|tool_call_end|>
-<|tool_calls_section_end|>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Tokyo"));
-        assert_eq!(normal_text, "Some text\n");
-    }
-
-    #[tokio::test]
-    async fn test_parse_kimi_multiple_tools() {
-        let parser = KimiK2Parser::new();
-        let input = r#"<|tool_calls_section_begin|>
-<|tool_call_begin|>functions.search:0<|tool_call_argument_begin|>{"query": "rust"}<|tool_call_end|>
-<|tool_call_begin|>functions.calculate:1<|tool_call_argument_begin|>{"expression": "2+2"}<|tool_call_end|>
-<|tool_calls_section_end|>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, "search");
-        assert_eq!(tools[1].function.name, "calculate");
-        assert_eq!(normal_text, "");
-    }
-
-    #[tokio::test]
-    async fn test_parse_kimi_with_whitespace() {
-        let parser = KimiK2Parser::new();
-        let input = r#"<|tool_calls_section_begin|>
-<|tool_call_begin|> functions.test:0 <|tool_call_argument_begin|> {"key": "value"} <|tool_call_end|>
-<|tool_calls_section_end|>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "test");
-        assert_eq!(normal_text, "");
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = KimiK2Parser::new();
-        assert!(parser.detect_format("<|tool_calls_section_begin|>"));
-        assert!(parser.detect_format("<|tool_call_begin|>"));
-        assert!(!parser.detect_format("plain text"));
-        assert!(!parser.detect_format("[TOOL_CALLS]"));
-    }
-}
