+++ sgl-router/src/tool_parser/parsers/llama_parser.rs
@@ -101,70 +101,3 @@ impl ToolParser for LlamaParser {
                 && (text.contains(r#""name""#) || text.contains(r#""function""#)))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_with_python_tag() {
-        let parser = LlamaParser::new();
-        let input = r#"<|python_tag|>{"name": "search", "arguments": {"query": "weather"}}"#;
-
-        let (normal_text, tool_calls) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tool_calls.len(), 1);
-        assert_eq!(tool_calls[0].function.name, "search");
-        assert!(tool_calls[0].function.arguments.contains("weather"));
-        assert_eq!(normal_text, ""); // Pure python_tag with JSON should have no normal text
-    }
-
-    #[tokio::test]
-    async fn test_parse_plain_json() {
-        let parser = LlamaParser::new();
-        let input = r#"{"name": "calculate", "arguments": {"x": 5, "y": 10}}"#;
-
-        let (normal_text, tool_calls) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tool_calls.len(), 1);
-        assert_eq!(tool_calls[0].function.name, "calculate");
-        assert_eq!(normal_text, ""); // Pure JSON should have no normal text
-    }
-
-    #[tokio::test]
-    async fn test_parse_with_text_before() {
-        let parser = LlamaParser::new();
-        let input = r#"Let me help you with that. <|python_tag|>{"name": "get_time", "arguments": {"timezone": "UTC"}}"#;
-
-        let (normal_text, tool_calls) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tool_calls.len(), 1);
-        assert_eq!(tool_calls[0].function.name, "get_time");
-        assert_eq!(normal_text, "Let me help you with that. ");
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = LlamaParser::new();
-
-        assert!(parser.detect_format(r#"<|python_tag|>{"name": "test"}"#));
-        assert!(parser.detect_format(r#"{"name": "test", "arguments": {}}"#));
-        assert!(!parser.detect_format("plain text"));
-        assert!(!parser.detect_format(r#"{"key": "value"}"#)); // No name field
-    }
-
-    #[tokio::test]
-    async fn test_single_call_with_semicolon() {
-        let parser = LlamaParser::new();
-        // Note: Llama 3.2 doesn't handle multiple calls well
-        let input = r#"<|python_tag|>{"name": "func1", "arguments": {"x": 1}};"#;
-
-        let (_normal_text, tool_calls) = parser.parse_complete(input).await.unwrap();
-
-        // We expect this to either parse the first JSON object or fail gracefully
-        // Since the semicolon makes it invalid JSON, it will likely return empty
-        // This is acceptable as Llama 3.2 doesn't reliably support parallel calls
-
-        // If it parses anything, it should be func1
-        if !tool_calls.is_empty() {
-            assert_eq!(tool_calls[0].function.name, "func1");
-        }
-    }
-}
