+++ sgl-router/src/tool_parser/parsers/qwen_parser.rs
@@ -291,136 +291,3 @@ impl ToolParser for QwenParser {
         true
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_qwen_format() {
-        let parser = QwenParser::new();
-        let input = r#"<tool_call>
-{"name": "get_weather", "arguments": {"location": "Beijing", "units": "celsius"}}
-</tool_call>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Beijing"));
-        assert_eq!(normal_text, ""); // Pure tool call, no normal text
-    }
-
-    #[tokio::test]
-    async fn test_parse_multiple_tools() {
-        let parser = QwenParser::new();
-        let input = r#"<tool_call>
-{"name": "search", "arguments": {"query": "rust programming"}}
-</tool_call>
-<tool_call>
-{"name": "calculate", "arguments": {"expression": "2 + 2"}}
-</tool_call>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, "search");
-        assert_eq!(tools[1].function.name, "calculate");
-        assert_eq!(normal_text, ""); // Pure tool calls, no normal text
-    }
-
-    #[tokio::test]
-    async fn test_with_normal_text() {
-        let parser = QwenParser::new();
-        let input = r#"Let me help you with that.
-<tool_call>
-{"name": "get_info", "arguments": {"topic": "Rust"}}
-</tool_call>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_info");
-        assert_eq!(normal_text, "Let me help you with that.\n");
-    }
-
-    #[tokio::test]
-    async fn test_nested_json_structures() {
-        let parser = QwenParser::new();
-        let input = r#"<tool_call>
-{
-    "name": "process_data",
-    "arguments": {
-        "data": {
-            "nested": {
-                "array": [1, 2, 3],
-                "object": {"key": "value"}
-            }
-        }
-    }
-}
-</tool_call>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "process_data");
-        assert!(tools[0].function.arguments.contains("nested"));
-        assert_eq!(normal_text, ""); // Pure tool call, no normal text
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = QwenParser::new();
-
-        assert!(parser.detect_format(
-            r#"<tool_call>
-{"name": "test", "arguments": {}}
-</tool_call>"#
-        ));
-
-        assert!(parser.detect_format(
-            r#"Text before <tool_call>
-{"name": "test", "arguments": {}}
-</tool_call> text after"#
-        ));
-
-        assert!(!parser.detect_format(r#"{"name": "test", "arguments": {}}"#));
-        assert!(!parser.detect_format("plain text"));
-
-        // Partial format should still be detected
-        assert!(parser.detect_format("<tool_call>"));
-    }
-
-    #[tokio::test]
-    async fn test_streaming_partial() {
-        let parser = QwenParser::new();
-        let mut state = ParseState::new();
-
-        // Simulate streaming chunks
-        let chunks = vec![
-            "<tool_call>\n",
-            r#"{"name": "search","#,
-            r#" "arguments": {"query":"#,
-            r#" "rust"}}"#,
-            "\n</tool_call>",
-        ];
-
-        let mut found_name = false;
-        let mut found_complete = false;
-
-        for chunk in chunks {
-            let result = parser.parse_incremental(chunk, &mut state).await.unwrap();
-
-            match result {
-                StreamResult::ToolName { name, .. } => {
-                    assert_eq!(name, "search");
-                    found_name = true;
-                }
-                StreamResult::ToolComplete(tool) => {
-                    assert_eq!(tool.function.name, "search");
-                    found_complete = true;
-                }
-                _ => {}
-            }
-        }
-
-        assert!(found_name || found_complete); // At least one should be found
-    }
-}
