+++ sgl-router/src/tool_parser/parsers/step3_parser.rs
@@ -284,77 +284,3 @@ impl ToolParser for Step3Parser {
         self.has_tool_markers(text)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_step3_single_tool() {
-        let parser = Step3Parser::new();
-        let input = r#"Some text
-<｜tool_calls_begin｜>
-<｜tool_call_begin｜>function<｜tool_sep｜><steptml:invoke name="get_weather">
-<steptml:parameter name="location">Tokyo</steptml:parameter>
-<steptml:parameter name="units">celsius</steptml:parameter>
-</steptml:invoke><｜tool_call_end｜>
-<｜tool_calls_end｜>"#;
-
-        let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "get_weather");
-        assert!(tools[0].function.arguments.contains("Tokyo"));
-        assert!(tools[0].function.arguments.contains("celsius"));
-        assert_eq!(normal_text, "Some text\n")
-    }
-
-    #[tokio::test]
-    async fn test_parse_step3_multiple_tools() {
-        let parser = Step3Parser::new();
-        let input = r#"<｜tool_calls_begin｜>
-<｜tool_call_begin｜>function<｜tool_sep｜><steptml:invoke name="search">
-<steptml:parameter name="query">rust programming</steptml:parameter>
-</steptml:invoke><｜tool_call_end｜>
-<｜tool_call_begin｜>function<｜tool_sep｜><steptml:invoke name="calculate">
-<steptml:parameter name="expression">2 + 2</steptml:parameter>
-</steptml:invoke><｜tool_call_end｜>
-<｜tool_calls_end｜>"#;
-
-        let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, "search");
-        assert_eq!(tools[1].function.name, "calculate");
-    }
-
-    #[tokio::test]
-    async fn test_parse_step3_mixed_types() {
-        let parser = Step3Parser::new();
-        let input = r#"<｜tool_calls_begin｜>
-<｜tool_call_begin｜>function<｜tool_sep｜><steptml:invoke name="process_data">
-<steptml:parameter name="count">42</steptml:parameter>
-<steptml:parameter name="active">true</steptml:parameter>
-<steptml:parameter name="rate">1.5</steptml:parameter>
-<steptml:parameter name="name">test</steptml:parameter>
-</steptml:invoke><｜tool_call_end｜>
-<｜tool_calls_end｜>"#;
-
-        let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
-        assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, "process_data");
-
-        // Parse arguments to check types
-        let args: serde_json::Value = serde_json::from_str(&tools[0].function.arguments).unwrap();
-        assert_eq!(args["count"], 42);
-        assert_eq!(args["active"], true);
-        assert_eq!(args["rate"], 1.5);
-        assert_eq!(args["name"], "test");
-    }
-
-    #[test]
-    fn test_detect_format() {
-        let parser = Step3Parser::new();
-        assert!(parser.detect_format("<｜tool_calls_begin｜>"));
-        assert!(!parser.detect_format("plain text"));
-        assert!(!parser.detect_format("[TOOL_CALLS]"));
-    }
-}
