+++ sgl-router/tests/tool_parser_kimik2.rs
@@ -12,8 +12,9 @@ async fn test_kimik2_complete_parsing() {
 <|tool_calls_section_end|>
 The weather in Tokyo is..."#;
 
-    let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
+    let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
     assert_eq!(tools.len(), 1);
+    assert_eq!(normal_text, "Let me help you with that.\n");
     assert_eq!(tools[0].function.name, "get_weather");
 
     let args: serde_json::Value = serde_json::from_str(&tools[0].function.arguments).unwrap();
@@ -30,8 +31,9 @@ async fn test_kimik2_multiple_tools() {
 <|tool_call_begin|>functions.translate:1<|tool_call_argument_begin|>{"text": "Hello", "to": "ja"}<|tool_call_end|>
 <|tool_calls_section_end|>"#;
 
-    let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
+    let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
     assert_eq!(tools.len(), 2);
+    assert_eq!(normal_text, "");
     assert_eq!(tools[0].function.name, "search");
     assert_eq!(tools[1].function.name, "translate");
 }
@@ -44,8 +46,9 @@ async fn test_kimik2_with_whitespace() {
 <|tool_call_begin|> functions.test:0 <|tool_call_argument_begin|> {"key": "value", "num": 42} <|tool_call_end|>
 <|tool_calls_section_end|>"#;
 
-    let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
+    let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
     assert_eq!(tools.len(), 1);
+    assert_eq!(normal_text, "");
     assert_eq!(tools[0].function.name, "test");
 
     let args: serde_json::Value = serde_json::from_str(&tools[0].function.arguments).unwrap();
@@ -117,8 +120,9 @@ async fn test_kimik2_sequential_indices() {
 <|tool_call_begin|>functions.third:2<|tool_call_argument_begin|>{"param": "c"}<|tool_call_end|>
 <|tool_calls_section_end|>"#;
 
-    let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
+    let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
     assert_eq!(tools.len(), 3);
+    assert_eq!(normal_text, "");
     assert_eq!(tools[0].function.name, "first");
     assert_eq!(tools[1].function.name, "second");
     assert_eq!(tools[2].function.name, "third");
@@ -134,12 +138,12 @@ async fn test_function_index_extraction() {
 <|tool_call_begin|>functions.calc:1<|tool_call_argument_begin|>{"x": 10}<|tool_call_end|>
 <|tool_calls_section_end|>"#;
 
-    let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
+    let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
     assert_eq!(tools.len(), 2);
+    assert_eq!(normal_text, "Text before tool calls.\n");
     assert_eq!(tools[0].function.name, "search");
     assert_eq!(tools[1].function.name, "calc");
     // TODO: Verify indices are preserved: 0 and 1
-    // TODO: Verify normal text = "Text before tool calls."
 }
 
 #[tokio::test]
