+++ sgl-router/tests/tool_parser_pythonic.rs
@@ -165,8 +165,9 @@ async fn test_pythonic_real_world_llama4() {
 
 These functions will provide the information you need."#;
 
-    let (_normal_text, tools) = parser.parse_complete(input).await.unwrap();
+    let (normal_text, tools) = parser.parse_complete(input).await.unwrap();
     assert_eq!(tools.len(), 3);
+    assert_eq!(normal_text, "I'll help you with multiple tasks. Let me search for information and perform calculations.\n\n\n\nThese functions will provide the information you need.");
     assert_eq!(tools[0].function.name, "web_search");
     assert_eq!(tools[1].function.name, "calculate");
     assert_eq!(tools[2].function.name, "get_weather");
